Added a cast to avoid a compiler warning

+++ modules/juce_audio_plugin_client/VST/juce_VST_Wrapper.cpp
@@ -446,7 +446,7 @@ public:
                         {
                             if (outputs[j] == chan)
                             {
-                                chan = new FloatType [blockSize * 2];
+                                chan = new FloatType [(size_t) blockSize * 2];
                                 tmpBuffers.tempChannels.set (i, chan);
                                 break;
                             }
